… tests/images/svg/Ghostscript_Tiger.png