tiger before:
`svg decode ........................ 38.983 ms     40.207 ms    ±0.562   x124`

tiger after:
`svg decode ........................ 33.804 ms     35.112 ms    ±0.556   x142`